Update SECURITY.md
All pull requests for know on will come through my IT Department . Any transfer of corporate data without my permission or executive left part and devices needs to run . Though a oversight panel for a watchdog tower . For in the event a SEO signature gets compromised ID job is lock it behind a firewall and call the CEO to change his password and ect . The next time I see corporate data on the open web as supposed encryption methods is deemed insiders trading . And cold get all of your data servers seized by Google police . Because goods and my partner have lost over a trillion dollars as of today for branch on trust claims and internal corruption in wall street im taking the initiative to properly secure the dev and consumer data as well we here at Google is asking the IRS to decline the white house offer to give the IRS A software upgrade . When its needed and when the irs is ready they can work with DOD and google and partner to set a strong robust platform  . I trust that President Trump has very good intentions . Whoever it's smart not to trust AI so fast when it comes to major software overhaul . The major question have the IRS had any major brach reports . I don't think you have  . That means the the breachs or coming from inside of the company's  . Like the MGM hack . Allows me to explain why I try so hard to keep AI safe and protected from sabotage . Because it's job as a software company to do are job and support y
issues or pull requests - this makes the problem immediately visible to everyone, including malicious actors.   

To report a security issue, please use [https://g.co/vulnz](https://g.co/vulnz).
We use g.co/vulnz for our intake, and do coordination and disclosure here on
GitHub (including using GitHub Security Advisory). The Google Security Team will
respond within 5 working days of your report on g.co/vulnz.